+++ frontend/src/app/home/components/PerformanceReview.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import dynamic from 'next/dynamic';
-import { generateSelfReview } from '@services/review';
+import { generateReview } from '@services/review';
 import { isBlankObject } from '@utils/object';
 import { TextButton } from '@components/ui/button';
 import { useFlashMessage } from '@components/ui/flash-messages';
@@ -12,27 +12,35 @@ const AudioInputComponent = dynamic(() => import('./AudioInput'), {
   ssr: false,
 });
 
-interface SelfReviewProps {
+interface PerformanceReviewProps {
   paramsWhenKeysNeeded: {
     userApiKey?: string;
     // Add other properties as needed
   };
   onReviewResultsReceived: (
-    self_review: Array<{ question: string; answer: string }>
+    review: Array<{ question: string; answer: string }>
   ) => void;
 }
 
-const SelfReview = ({
+interface FormState {
+  yourRole: string;
+  candidateRole: string;
+  perfQuestion: string;
+  yourReview: string;
+}
+
+const PerformanceReview = ({
   paramsWhenKeysNeeded,
   onReviewResultsReceived,
-}: SelfReviewProps) => {
+}: PerformanceReviewProps) => {
   const { addInfoMessage, addFailureMessage } = useFlashMessage();
   const { accessToken } = useAppContext();
 
-  const [formState, setFormState] = useState({
-    textDump: '',
-    questions: '',
-    instructions: '',
+  const [formState, setFormState] = useState<FormState>({
+    yourRole: '',
+    candidateRole: '',
+    perfQuestion: '',
+    yourReview: '',
   });
   const [transcription, setTranscription] = useState<string>('');
   const [isLoading, setIsLoading] = useState(false);
@@ -51,15 +59,15 @@ const SelfReview = ({
     ) {
       throw new Error('Please enter your API key.');
     }
-    if (!formState.questions) {
-      throw new Error('Please provide both the text dump and questions.');
+    if (!formState.yourRole || !formState.candidateRole) {
+      throw new Error('Please fill in all required fields.');
     }
-    if (!formState.textDump && !transcription) {
-      throw new Error('Please provide either a text dump or audio input.');
+    if (!formState.yourReview && !transcription) {
+      throw new Error('Please provide either a review dump or audio input.');
     }
   };
 
-  const handleGenerateSelfReview = async () => {
+  const handleGenerateReview = async () => {
     if (!accessToken) {
       addInfoMessage({ message: 'Please login to generate review' });
       return;
@@ -71,25 +79,22 @@ const SelfReview = ({
       validateForm();
 
       const requestData = {
-        text_dump: formState.textDump + transcription,
-        questions: formState.questions
-          .split('\n')
-          .map((q) => q.trim())
-          .filter(Boolean),
-        instructions: formState.instructions || null,
+        your_role: formState.yourRole,
+        candidate_role: formState.candidateRole,
+        perf_question: formState.perfQuestion,
+        your_review: `${formState.yourReview} ${transcription}`.trim(),
         is_paid: false,
         ...paramsWhenKeysNeeded,
       };
 
-      const response = (await generateSelfReview(requestData)) as {
-        self_review: Array<{ question: string; answer: string }>;
+      const response = (await generateReview(requestData)) as {
+        review: Array<{ question: string; answer: string }>;
       };
-      onReviewResultsReceived(response.self_review);
+      onReviewResultsReceived(response.review);
     } catch (error: any) {
-      console.error('Error generating self-review:', error);
-
+      console.error('Error generating review:', error);
       const errMsg =
-        error?.message || 'An error occurred while generating the self review.';
+        error?.message || 'An error occurred while generating the review.';
       addFailureMessage({ message: errMsg, autoClose: false });
     } finally {
       setIsLoading(false);
@@ -97,29 +102,43 @@ const SelfReview = ({
   };
 
   return (
-    <section className="relative isolate px-6 py-4 lg:py-8 lg:px-8 widget-animate animate in-view">
-      <div className="flex flex-col shadow-2xl mx-auto max-w-5xl justify-between gap-8 border-secondary items-center rounded-xl border bg-zinc-800 p-6 sm:p-12 sm:px-12  dark:bg-zinc-900">
+    <div className="relative isolate w-full widget-animate animate in-view">
+      <div className="flex flex-col shadow-2xl mx-auto w-full justify-between gap-8 border-secondary items-center border bg-zinc-800 p-6 sm:p-12 sm:px-12  dark:bg-zinc-900">
         <div className="w-full">
           <label className="block text-milk text-sm font-medium mb-2">
-            Text Dump (information about your performance)
+            Your Role
           </label>
-          <textarea
-            value={formState.textDump}
-            onChange={updateFormField('textDump')}
-            placeholder="Please enter your inputs..."
+          <input
+            type="text"
+            value={formState.yourRole}
+            onChange={updateFormField('yourRole')}
+            placeholder="Please enter your role"
             className="w-full px-3 py-2 border rounded
               bg-background border-gray-700 text-gray-300
               placeholder:text-gray-600 hover:bg-neutral-800"
-            rows={4}
           />
         </div>
         <div className="w-full">
           <label className="block text-milk text-sm font-medium mb-2">
-            Questions to Answer in Self-Review (one per line)
+            Candidate Role
+          </label>
+          <input
+            type="text"
+            value={formState.candidateRole}
+            onChange={updateFormField('candidateRole')}
+            placeholder="Please enter candidate role"
+            className="w-full px-3 py-2 border rounded
+              bg-background border-gray-700 text-gray-300
+              placeholder:text-gray-600 hover:bg-neutral-800"
+          />
+        </div>
+        <div className="w-full">
+          <label className="block text-milk text-sm font-medium mb-2">
+            Performance Review Questions (one per line)
           </label>
           <textarea
-            value={formState.questions}
-            onChange={updateFormField('questions')}
+            value={formState.perfQuestion}
+            onChange={updateFormField('perfQuestion')}
             placeholder="Please enter questions..."
             className="w-full px-3 py-2 border rounded
               bg-background border-gray-700 text-gray-300
@@ -129,12 +148,12 @@ const SelfReview = ({
         </div>
         <div className="w-full">
           <label className="block text-milk text-sm font-medium mb-2">
-            Additional Instructions (optional)
+            Your Review
           </label>
           <textarea
-            value={formState.instructions}
-            onChange={updateFormField('instructions')}
-            placeholder="Please enter additional instructions..."
+            value={formState.yourReview}
+            onChange={updateFormField('yourReview')}
+            placeholder="Please enter your review..."
             className="w-full px-3 py-2 border rounded
               bg-background border-gray-700 text-gray-300
               placeholder:text-gray-600 hover:bg-neutral-800"
@@ -148,17 +167,17 @@ const SelfReview = ({
           />
         </div>
         <TextButton
-          onClick={handleGenerateSelfReview}
-          variant={`primary-self-review`}
+          onClick={handleGenerateReview}
+          variant={`primary-perf-review`}
           disabled={isLoading}
           className="font-semibold"
           startIcon={<SvgIcon svg={AiIcon} size='lg' />}
         >
-          {isLoading ? 'Generating...' : 'Generate Self Review'}
+          {isLoading ? 'Generating...' : 'Generate Performance Review'}
         </TextButton>
       </div>
-    </section>
+    </div>
   );
 };
 
-export default SelfReview;
+export default PerformanceReview;
