+++ frontend/src/app/home/components/SelfReview.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import dynamic from 'next/dynamic';
-import { generateReview } from '@services/review';
+import { generateSelfReview } from '@services/review';
 import { isBlankObject } from '@utils/object';
 import { TextButton } from '@components/ui/button';
 import { useFlashMessage } from '@components/ui/flash-messages';
@@ -12,35 +12,27 @@ const AudioInputComponent = dynamic(() => import('./AudioInput'), {
   ssr: false,
 });
 
-interface PerformanceReviewProps {
+interface SelfReviewProps {
   paramsWhenKeysNeeded: {
     userApiKey?: string;
     // Add other properties as needed
   };
   onReviewResultsReceived: (
-    review: Array<{ question: string; answer: string }>
+    self_review: Array<{ question: string; answer: string }>
   ) => void;
 }
 
-interface FormState {
-  yourRole: string;
-  candidateRole: string;
-  perfQuestion: string;
-  yourReview: string;
-}
-
-const PerformanceReview = ({
+const SelfReview = ({
   paramsWhenKeysNeeded,
   onReviewResultsReceived,
-}: PerformanceReviewProps) => {
+}: SelfReviewProps) => {
   const { addInfoMessage, addFailureMessage } = useFlashMessage();
   const { accessToken } = useAppContext();
 
-  const [formState, setFormState] = useState<FormState>({
-    yourRole: '',
-    candidateRole: '',
-    perfQuestion: '',
-    yourReview: '',
+  const [formState, setFormState] = useState({
+    textDump: '',
+    questions: '',
+    instructions: '',
   });
   const [transcription, setTranscription] = useState<string>('');
   const [isLoading, setIsLoading] = useState(false);
@@ -59,15 +51,15 @@ const PerformanceReview = ({
     ) {
       throw new Error('Please enter your API key.');
     }
-    if (!formState.yourRole || !formState.candidateRole) {
-      throw new Error('Please fill in all required fields.');
+    if (!formState.questions) {
+      throw new Error('Please provide both the text dump and questions.');
     }
-    if (!formState.yourReview && !transcription) {
-      throw new Error('Please provide either a review dump or audio input.');
+    if (!formState.textDump && !transcription) {
+      throw new Error('Please provide either a text dump or audio input.');
     }
   };
 
-  const handleGenerateReview = async () => {
+  const handleGenerateSelfReview = async () => {
     if (!accessToken) {
       addInfoMessage({ message: 'Please login to generate review' });
       return;
@@ -79,22 +71,25 @@ const PerformanceReview = ({
       validateForm();
 
       const requestData = {
-        your_role: formState.yourRole,
-        candidate_role: formState.candidateRole,
-        perf_question: formState.perfQuestion,
-        your_review: `${formState.yourReview} ${transcription}`.trim(),
+        text_dump: formState.textDump + transcription,
+        questions: formState.questions
+          .split('\n')
+          .map((q) => q.trim())
+          .filter(Boolean),
+        instructions: formState.instructions || null,
         is_paid: false,
         ...paramsWhenKeysNeeded,
       };
 
-      const response = (await generateReview(requestData)) as {
-        review: Array<{ question: string; answer: string }>;
+      const response = (await generateSelfReview(requestData)) as {
+        self_review: Array<{ question: string; answer: string }>;
       };
-      onReviewResultsReceived(response.review);
+      onReviewResultsReceived(response.self_review);
     } catch (error: any) {
-      console.error('Error generating review:', error);
+      console.error('Error generating self-review:', error);
+
       const errMsg =
-        error?.message || 'An error occurred while generating the review.';
+        error?.message || 'An error occurred while generating the self review.';
       addFailureMessage({ message: errMsg, autoClose: false });
     } finally {
       setIsLoading(false);
@@ -102,43 +97,29 @@ const PerformanceReview = ({
   };
 
   return (
-    <section className="relative isolate px-6 py-4 lg:py-8 lg:px-8 widget-animate animate in-view">
-      <div className="flex flex-col shadow-2xl mx-auto max-w-5xl justify-between gap-8 border-secondary items-center rounded-xl border bg-zinc-800 p-6 sm:p-12 sm:px-12  dark:bg-zinc-900">
+    <section className="relative isolate w-full widget-animate animate in-view">
+      <div className="flex flex-col shadow-2xl mx-auto w-full justify-between gap-8 border-secondary items-center border bg-zinc-800 p-6 sm:p-12 sm:px-12  dark:bg-zinc-900">
         <div className="w-full">
           <label className="block text-milk text-sm font-medium mb-2">
-            Your Role
-          </label>
-          <input
-            type="text"
-            value={formState.yourRole}
-            onChange={updateFormField('yourRole')}
-            placeholder="Please enter your role"
-            className="w-full px-3 py-2 border rounded
-              bg-background border-gray-700 text-gray-300
-              placeholder:text-gray-600 hover:bg-neutral-800"
-          />
-        </div>
-        <div className="w-full">
-          <label className="block text-milk text-sm font-medium mb-2">
-            Candidate Role
-          </label>
-          <input
-            type="text"
-            value={formState.candidateRole}
-            onChange={updateFormField('candidateRole')}
-            placeholder="Please enter candidate role"
-            className="w-full px-3 py-2 border rounded
-              bg-background border-gray-700 text-gray-300
-              placeholder:text-gray-600 hover:bg-neutral-800"
-          />
-        </div>
-        <div className="w-full">
-          <label className="block text-milk text-sm font-medium mb-2">
-            Performance Review Questions (one per line)
+            Text Dump (information about your performance)
           </label>
           <textarea
-            value={formState.perfQuestion}
-            onChange={updateFormField('perfQuestion')}
+            value={formState.textDump}
+            onChange={updateFormField('textDump')}
+            placeholder="Please enter your inputs..."
+            className="w-full px-3 py-2 border rounded
+              bg-background border-gray-700 text-gray-300
+              placeholder:text-gray-600 hover:bg-neutral-800"
+            rows={4}
+          />
+        </div>
+        <div className="w-full">
+          <label className="block text-milk text-sm font-medium mb-2">
+            Questions to Answer in Self-Review (one per line)
+          </label>
+          <textarea
+            value={formState.questions}
+            onChange={updateFormField('questions')}
             placeholder="Please enter questions..."
             className="w-full px-3 py-2 border rounded
               bg-background border-gray-700 text-gray-300
@@ -148,12 +129,12 @@ const PerformanceReview = ({
         </div>
         <div className="w-full">
           <label className="block text-milk text-sm font-medium mb-2">
-            Your Review
+            Additional Instructions (optional)
           </label>
           <textarea
-            value={formState.yourReview}
-            onChange={updateFormField('yourReview')}
-            placeholder="Please enter your review..."
+            value={formState.instructions}
+            onChange={updateFormField('instructions')}
+            placeholder="Please enter additional instructions..."
             className="w-full px-3 py-2 border rounded
               bg-background border-gray-700 text-gray-300
               placeholder:text-gray-600 hover:bg-neutral-800"
@@ -167,17 +148,17 @@ const PerformanceReview = ({
           />
         </div>
         <TextButton
-          onClick={handleGenerateReview}
-          variant={`primary-perf-review`}
+          onClick={handleGenerateSelfReview}
+          variant={`primary-self-review`}
           disabled={isLoading}
           className="font-semibold"
           startIcon={<SvgIcon svg={AiIcon} size='lg' />}
         >
-          {isLoading ? 'Generating...' : 'Generate Performance Review'}
+          {isLoading ? 'Generating...' : 'Generate Self Review'}
         </TextButton>
       </div>
     </section>
   );
 };
 
-export default PerformanceReview;
+export default SelfReview;
